Refactor entire project to CRApp-less
Using Webpack, the app is no longer reliant on Create React App.
