--- config-overrides/rewireEslint.js
@@ -1,3 +1,0 @@
-const rewireEslint = require("react-app-rewire-eslint");
-
-module.exports = (config, env) => rewireEslint(config, env);
