--- config-overrides/rewireHotLoader.js
@@ -1,3 +1,0 @@
-const rewireHotLoader = require("react-app-rewire-hot-loader");
-
-module.exports = (config, env) => rewireHotLoader(config, env);
