--- jsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2017",
-    "baseUrl": "./",
-    "paths": {
-      "@/*": ["src/*"]
-    }
-  },
-  "include": ["src"]
-}
