+++ package.json
@@ -2,86 +2,44 @@
   "name": "star-wars-generals",
   "version": "0.1.0",
   "private": true,
+  "main": "index.js",
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/preset-react"
+    ]
+  },
   "dependencies": {
-    "@material-ui/core": "^3.2.0",
-    "classnames": "^2.2.6",
-    "connected-react-router": "^4.5.0",
-    "prop-types": "^15.6.2",
-    "ramda": "^0.25.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-markdown": "^3.6.0",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^1.1.5",
-    "redux": "^3.7.2",
-    "redux-actions": "^2.4.0",
-    "redux-devtools-extension": "^2.13.5",
-    "redux-saga": "v1.0.0-beta.1"
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test"
+    "start": "webpack-dev-server --open --hot",
+    "build": "webpack"
   },
   "devDependencies": {
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.6.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-import-resolver-webpack": "^0.10.1",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-jest": "^21.25.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.9.1",
-    "eslint-plugin-redux-saga": "^0.9.0",
-    "jest-each": "^23.6.0",
-    "jest-enzyme": "^6.1.2",
-    "jest-fetch-mock": "^1.6.6",
-    "node-sass": "^4.9.2",
-    "prettier": "^1.14.3",
-    "react-app-rewire-eslint": "^0.2.3",
-    "react-app-rewire-hot-loader": "^1.0.3",
-    "react-app-rewired": "^1.6.2",
-    "react-hot-loader": "^4.3.11",
-    "redux-saga-test-plan": "v4.0.0-beta.1",
-    "sass-loader": "^7.0.3"
-  },
-  "jest": {
-    "resetMocks": true,
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}",
-      "!**/index.js",
-      "!src/registerServiceWorker.js",
-      "!src/SetupTests.js",
-      "!src/PropTypes/**"
-    ],
-    "coverageReporters": [
-      "json",
-      "lcov",
-      "text"
-    ],
-    "coverageThreshold": {
-      "global": {
-        "branches": 80,
-        "functions": 80,
-        "lines": 80,
-        "statements": 80
-      }
-    }
-  },
-  "browserslist": {
-    "development": [
-      "last 2 chrome versions",
-      "last 2 firefox versions",
-      "last 2 edge versions"
-    ],
-    "production": [
-      ">0.25%",
-      "not op_mini all",
-      "ie 11"
-    ]
+    "@babel/core": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.5",
+    "css-loader": "^2.1.1",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-loader": "^2.1.2",
+    "eslint-plugin-flowtype": "^3.4.2",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jest": "^22.4.1",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.12.4",
+    "file-loader": "^3.0.1",
+    "html-webpack-plugin": "^3.2.0",
+    "prettier": "^1.16.4",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0",
+    "webpack-dev-server": "^3.2.1"
   }
 }
