+++ src/components/BattleView/Grave/Grave.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import Card from "@/components/card";
+import Card from "../../Card";
 
 import "./Grave.css";
