+++ src/components/MainMenu/MainMenu.jsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Link } from "react-router-dom";
 
-import Logo from "@/assets/images/logo.png";
+import Logo from "../../assets/images/logo.png";
 
 import "./MainMenu.css";
