+++ src/components/TrainingView/TrainingView.jsx
@@ -4,6 +4,10 @@ import "./TrainingView.css";
 
 export default class TrainingView extends React.PureComponent {
   render() {
-    return <div>Training View</div>;
+    return (
+      <div>
+        <h1>Training View</h1>
+      </div>
+    );
   }
 }
