+++ webpack.config.js
@@ -1,0 +1,31 @@
+const path = require("path");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+
+module.exports = {
+  entry: "./src/index.js",
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "index_bundle.js"
+  },
+  module: {
+    rules: [
+      { test: /\.(js|jsx)$/, exclude: /node_modules/, use: "babel-loader" },
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: ["babel-loader", "eslint-loader"]
+      },
+      { test: /\.css$/, use: ["style-loader", "css-loader"] },
+      { test: /\.(png|svg|jpg|gif)$/, use: "file-loader" }
+    ]
+  },
+  mode: "development",
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: "public/index.html"
+    })
+  ],
+  resolve: {
+    extensions: [".js", ".json", ".jsx", ".png"]
+  }
+};
